Seems to work, but can explode if sensor is set to none but left plugged in.
Can't free the static dummy sensor memory, so fixed that.
Had to increase the ES2 on/off interval to get serial from it.
Works most of the time, but it is possible to start sending gibberish messages that cause the phone to crash.
This seems to occur when you leave a sensor plugged in after it was working and then tell the firmware it is dummy/none.
The sensor (or another sensor) may then start sending gibberish very very quickly to the phone, causing it to crash.
For example, this occurred when S1 started as DO and S3 started as ES2, and then S1 is set to dummy.
This caused S3 to send ridiculous messages much faster than it should.
The JSON parsing in the phone is protected by try/catch that works when gibberish is sent that normal rates,
  but reading.type.toString() will cause an uncaught null pointer exception if it is sent at extreme rates.