Fixed gibberish issue by using dummy serial handler callback.
The previous commit caused gibberish to appear on other sensor channels when a
real sensor was set to dummy, e.g.

DO, dummy, ES2 --> dummy, dummy, ES2

could cause the ES2 channel to produce gibberish extremely quickly,
overwhelming the phone and crashing it.

The serial handlers were stored in a separate global array.
Setting the sensor to a dummy did not change what was already set up as a callback reference,
and this probably caused the undefined behavior.

Now, when a sensor is set back to a dummy, its corresponding serial handler is as well.