Finally plugged in a FrSky X8R and Taranis pair, but something is wrong.
The signals do not make sense. They are in the correct range of values,
but they are either unstable, or they are certainly wrong values.
For example, if the trim is set to neutral, and the thrust is all the way
down, the signal is 940 (rather than < 200) but if you move the stick up
just two clicks or move the trim up one click, the signal jumps to a value
that makes sense. The resting position of another signal should be ~1000 but
is 1507! All these things make it unusable! Need to fix ASAP.