Can now turn on override and run motors, but receiver can ignore transmitter after boat changes state.
Works best with no phone in the boat at all.
If the receiver doesn't ignore the transmitter, it functions as intended.
But sometimes (especially when the boat changes USB state) the reciever seems to ignore the transmitter.
I don't know why this is. Sometimes you have to power cycle the arduino several times before it works again.