Cleaned up PWM. Now matches SBUS where applicable.
Could further combine the two update methods under a nonvirtual RC::update.
But this is complicated by the nature of the PWM override.
In SBUS, the override jumps from minimum to maximum.
In PWM, the override jumps above and below the midpoint.
I'm sure there is a way to handle this, but isn't really necessary.